// Source/Web/SS.StudentStore.UI/src/app/model/product/productresponse.ts
export class ProductResponse {
    productId: number;
    productTypeId: number;
    title : string;
    name: string;
    description: string;
    productTypeName: string
    price: number;
    categoryId : number;
    categoryName : string;
    subCategoryId : number;
    subCategoryName : string;
    gradeId : number;
    gradeName : string;
    sectionId : number;
    sectionName : string;
    brandId : number;
    brandName : string;
    shortPictureUrl: string;
    pictureUrl: string;
    productManufacturer: string;
    productManufacturerPicture: string;
    documentUrl: string;
}